Fixed VTK XML writer: faces and face offsets need 32-bit ints, not 8-bit!
Former-commit-id: 6a81ae046e5c1ef6d4661f2cdb6d0209e9ebc440

+++ io/vtk_plot_io.c
@@ -318,7 +318,7 @@ static void vtk_plot_write_asci_datasets(void* context, void* f, io_dataset_t** 
     end_element(writer, "DataArray");
 
     start_element(writer, "DataArray");
-    write_attribute(writer, "type", "UInt8");
+    write_attribute(writer, "type", "UInt32");
     write_attribute(writer, "Name", "faces");
     write_attribute(writer, "format", "ascii");
 
@@ -358,7 +358,7 @@ static void vtk_plot_write_asci_datasets(void* context, void* f, io_dataset_t** 
     end_element(writer, "DataArray");
 
     start_element(writer, "DataArray");
-    write_attribute(writer, "type", "UInt8");
+    write_attribute(writer, "type", "UInt32");
     write_attribute(writer, "Name", "faceoffsets");
     write_attribute(writer, "format", "ascii");
